The main functionality of this is project is to add driver details, transport details and we can update and delete the information of both.
So basically my project name is transport management system there we can add the drivers information and transport information, In driver information we can add the drivers details like their name, phone, license number address, and in transport details we can add the transport name, type, number, route etc. So in frontend I have used the html, css and basic javascript for validation and in backend I used the servlet, jdbc and for database I've used the mysql. So main functionality of this is project is to add driver details, transport details and we can update and delete the information of both.